Extract files related scripts to files directory
