+++ scripts/files/paths.ts
